server: Restructure nats connection
Move nats connection internal to the ChatServer struct.  The idea of
using channels was to decouple the nats dependency from the chat server,
allowing for easier testing and the ability to swap out/skip NATS
entirely.  This is still worthwhile, but should be accomplished
differently.  And this is just a quick project to work with
terraform/k8s.. So don't over-think it.